Move middleware to a middleware folder
