+++ packages/meditrak-server/src/apiV2/middleware/auth/index.js
@@ -3,4 +3,4 @@
  * Copyright (c) 2017 Beyond Essential Systems Pty Ltd
  */
 
-export { authenticationMiddleware } from './middleware';
+export { authenticationMiddleware } from './authenticationMiddleware';
