+++ packages/meditrak-server/src/apiV2/middleware/catchAsyncErrors.js
@@ -1,0 +1,17 @@
+/**
+ * Tupaia
+ * Copyright (c) 2017 - 2021 Beyond Essential Systems Pty Ltd
+ */
+
+/**
+ * All routes will be wrapped with an error catcher that simply passes the error to the next()
+ * function, causing error handling middleware to be fired. Otherwise, async errors will be
+ * swallowed.
+ */
+export const catchAsyncErrors = routeHandler => (res, req, next) => {
+  const returnValue = routeHandler(res, req, next);
+  if (returnValue && returnValue.catch) {
+    // Return value is a Promise, add an error handler
+    returnValue.catch(error => next(error));
+  }
+};
