+++ packages/meditrak-server/src/apiV2/middleware/handleError.js
@@ -1,0 +1,21 @@
+/**
+ * Tupaia
+ * Copyright (c) 2017 - 2021 Beyond Essential Systems Pty Ltd
+ */
+import { InternalServerError } from '@tupaia/utils';
+
+export const handleError = (err, req, res, next) => {
+  const { database, apiRequestLogId } = req;
+  let error = err;
+  if (!error.respond) {
+    error = new InternalServerError(err);
+  }
+  if (database) {
+    database.create('error_log', {
+      message: error.message,
+      type: error.constructor.name,
+      api_request_log_id: apiRequestLogId,
+    });
+  }
+  error.respond(res);
+};
